+++ tests/unit/flush/flush-common.h
@@ -1,5 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /* Copyright 2020-2021, Intel Corporation */
+/* Copyright (c) 2023 Fujitsu Limited */
 
 /*
  * flush-common.h -- header of the common part of unit tests
@@ -10,6 +11,7 @@
 #define FLUSH_COMMON_H 1
 
 #include "mocks-stdlib.h"
+#include "mocks-ibverbs.h"
 
 #define MOCK_RPMA_MR_REMOTE	(struct rpma_mr_remote *)0xC412
 #define MOCK_RPMA_MR_LOCAL	(struct rpma_mr_local *)0xC411
@@ -20,14 +22,20 @@
 #define MOCK_RAW_LEN		8
 
 /*
- * All the resources used between setup__flush_new and teardown__flush_delete.
+ * All the resources used between setup__{apm, native}_flush_new and
+ * teardown__{apm, native}_flush_delete.
  */
 struct flush_test_state {
 	struct rpma_flush *flush;
 	struct mmap_args allocated_raw;
 };
 
-int setup__flush_new(void **fstate_ptr);
-int teardown__flush_delete(void **fstate_ptr);
+int setup__apm_flush_new(void **fstate_ptr);
+int teardown__apm_flush_delete(void **fstate_ptr);
+#ifdef NATIVE_FLUSH_SUPPORTED
+int setup__native_flush_new(void **fstate_ptr);
+int teardown__native_flush_delete(void **fstate_ptr);
+#endif
+int group_setup_flush_common(void **unused);
 
 #endif /* FLUSH_COMMON_H */